Refactor getTopmostReportParam to getTopmostReportParams
