+++ src/libs/Navigation/helpers/getTopmostReportParams.ts
@@ -1,22 +1,20 @@
 import type {NavigationState, PartialState} from '@react-navigation/native';
-import type {RootStackParamList} from '@libs/Navigation/types';
+import type {ReportsSplitNavigatorParamList, RootStackParamList} from '@libs/Navigation/types';
 import NAVIGATORS from '@src/NAVIGATORS';
 import SCREENS from '@src/SCREENS';
 
 // This function is in a separate file than Navigation.ts to avoid cyclic dependency.
 
 /**
- * Find the last visited report screen in the navigation state and get its specific param (id or action id).
+ * Find the last visited report screen in the navigation state and get its params.
  *
  * @param state - The react-navigation state
- * @param reportParam - param to get from the report route params
  * @returns - It's possible that there is no report screen
  */
 
 type State = NavigationState | NavigationState<RootStackParamList> | PartialState<NavigationState>;
-type ReportParam = 'reportID' | 'reportActionID';
 
-function getTopmostReportParam(state: State, reportParam: ReportParam): string | undefined {
+function getTopmostReportParams(state: State): ReportsSplitNavigatorParamList[typeof SCREENS.REPORT] | undefined {
     if (!state) {
         return;
     }
@@ -33,9 +31,7 @@ function getTopmostReportParam(state: State, reportParam: ReportParam): string |
         return;
     }
 
-    const topmostReportParams = topmostReport?.params as Record<string, string>;
-
-    return topmostReportParams?.[reportParam];
+    return topmostReport?.params as ReportsSplitNavigatorParamList[typeof SCREENS.REPORT];
 }
 
-export default getTopmostReportParam;
+export default getTopmostReportParams;
